feature to save user_agents for corresponding proxy urls so that sessions appear like the same browser/ip every time FG restarts.
